sagemaker_source/requirements.txt
transformers
torchaudio
torchvision
imageio-ffmpeg
opencv-python-headless
scipy
pandas
scikit-learn
numpy==1.26.4
